Laravel Web Todolist 08 Todo Test

+++ 11-laravel-web-todolist/app/Services/Impl/TodoListServiceImpl.php
@@ -38,6 +38,8 @@ class TodoListServiceImpl implements TodoListService
     public function removeTodo(string $todoId)
     {
         $todo = Todo::find($todoId);
-        $todo->delete();
+
+        if ($todo != null)
+            $todo->delete();
     }
 }
